Support CCID 1.10 and fix pts and parameter setting in the proper order
By: Chaskiel Grundman

I have apparently been sitting on a patch that does pts and parameter setting in the proper order for most of a year. (the problem you have seems to be that the existing code changes the baud rate before doing pts, so the card responds at the old baud rate) I do not know why I did not submit this patch earlier. I do know that I have not really tested it recently



git-svn-id: http://www.opensc-project.org/svn/openct/trunk@1037 4cd62497-39f2-0310-b04f-c6e0f9629ffd