Trigger signal when USB device disconnect
This enable safe detection of device removal and allow USB drivers
to stop polling device, even if they do not have interrupt interface.

As OpenCT uses /dev/bus/usb namespace if exists, and this feature worked
only on usbfs prior to linux-2.6.28, it will not work with older kernels.

linux-2.6.28 has a bug that cause the signal not to be sent if kernel
is configured without CONFIG_USB_DEVICEFS, see [1].

linux-2.6.29 is expected to solve this issue.

Even if ccid driver works correctly, as it uses interrupt endpoint,
by default polling will be forced.

[1] http://marc.info/?t=123075926900001&r=1&w=2


git-svn-id: http://www.opensc-project.org/svn/openct/branches/alonbl@1100 4cd62497-39f2-0310-b04f-c6e0f9629ffd